Uniformize DAQ_Move and DAQ_Viewer control modules
Add _hw_settings_name and _ui_init_attr class attributes to
ParameterControlModule so subclasses declare their settings group name
and UI init property once, removing scattered string literals.

Consolidate into the base class:
- quit_fun: common deinit → quit_signal → _quit_cleanup() hook → ui.close
- _quit_cleanup: no-op default; DAQ_Viewer overrides to close its LCD widget
- _close_hardware: emits 'close' to hardware thread, resets UI init state,
  disconnects LECO; replaces duplicated try/except/finally in both modules
- param_deleted: identical in both modules except for _hw_settings_name
- apply_controller_parameters: replace if/elif on module_type with _hw_settings_name
- process_leco_commands GET_SETTINGS: replace class-name heuristic with _hw_settings_name
